refactor dataset utils
